opencompass_test/requirements.txt
# OpenCompass evaluation dependencies
# Based on conda Rmoe environment (pip freeze)

# PyTorch CUDA index (required for torch+cu124 builds)
--extra-index-url https://download.pytorch.org/whl/cu124

# Core evaluation framework
opencompass==0.4.2

# Deep learning
torch==2.6.0+cu124
torchvision==0.21.0+cu124

# HuggingFace ecosystem
transformers==4.51.3
accelerate==1.4.0
datasets==3.5.0
huggingface_hub==0.36.2
safetensors==0.5.3
tokenizers==0.21.1
sentencepiece==0.2.0
tiktoken==0.9.0
peft==0.18.1

# Evaluation metrics
rouge==1.0.1
rouge-chinese==1.0.3
rouge_score==0.1.2
sacrebleu==2.5.1

# NLP utilities
jieba==0.42.1
fuzzywuzzy==0.18.0
nltk==3.9.1
langdetect==1.0.9

# Scientific computing
numpy==1.26.4
scipy==1.15.2
pandas==1.5.3

# Display and utilities
prettytable==3.16.0
tabulate==0.9.0
tqdm==4.67.1
psutil==7.0.0
mmengine-lite==0.10.7
setuptools<70

# Flash attention (optional, GPU only, requires CUDA build)
# flash-attn==2.7.4.post1
